+++ CHANGELOG.md
@@ -11,8 +11,6 @@ _This release is scheduled to be released on 2025-01-01._
 
 ### Added
 
-- [linter] re-added `eslint-plugin-import`now that it supports ESLint v9
-
 ### Removed
 
 - [tests] Removed node-pty and drivelist from rebuilded test (#3575)
@@ -29,6 +27,7 @@ _This release is scheduled to be released on 2025-01-01._
 - [updatenotification] Fix pm2 using detection when pm2 script is in MagicMirror root folder (#3576)
 - [core] Fix loading node_helper of modules: avoid black screen, display errors and continue loading with next module (#3578)
 - [weather] changed default value for weatherEndpoint of provider openweathermap to "/onecall" (#3574)
+- [calendar] - update to resolve issues #3098 #3144 #3351 #3422 #3443 #3467 #3537 related to timezone changes
 
 ## [2.29.0] - 2024-10-01